ai-test-architect/requirements.txt
robotframework>=7.0
langchain
langchain-openai
python-dotenv
streamlit
# --- NEW ---
fastapi
uvicorn
requests
pydantic